+++ neighbor_list.h
@@ -23,17 +23,17 @@ __global__ void ksScanInc (int *X, int *Y, int InputSize);
 __global__ void sumIt (int *Y, int *S, int InputSize);
 
 __global__ void copyElements(int *dev_index, int *dev_value, int *dev_output, int *dev_result, int N);
-__global__ void copyElements(double *dev_index, int *dev_value, int *dev_output, double *dev_result, int N);
+__global__ void copyElements(float *dev_index, int *dev_value, int *dev_output, float *dev_result, int N);
 
 void calculate_array_native(int boxl, int N);
-__global__ void array_native_kernel(int *dev_ibead_lj_nat, int *dev_jbead_lj_nat, int *dev_itype_lj_nat, int *dev_jtype_lj_nat, double3 *dev_unc_pos, double *dev_lj_nat_pdb_dist, int *dev_value, int boxl, int N);
+__global__ void array_native_kernel(int *dev_ibead_lj_nat, int *dev_jbead_lj_nat, int *dev_itype_lj_nat, int *dev_jtype_lj_nat, float3 *dev_unc_pos, float *dev_lj_nat_pdb_dist, int *dev_value, int boxl, int N);
 
 void calculate_array_non_native(int boxl, int N);
-__global__ void array_non_native_kernel(int *dev_ibead_lj_non_nat, int *dev_jbead_lj_non_nat, int *dev_itype_lj_non_nat, int *dev_jtype_lj_non_nat, double3 *dev_unc_pos, int *dev_value, int boxl, int N);
+__global__ void array_non_native_kernel(int *dev_ibead_lj_non_nat, int *dev_jbead_lj_non_nat, int *dev_itype_lj_non_nat, int *dev_jtype_lj_non_nat, float3 *dev_unc_pos, int *dev_value, int boxl, int N);
 
 void hier_ks_scan(int *dev_X, int *dev_Y, int N, int re);
 int compact(int *index, int *value, int N, int *&result);
-int compact(double *index, int *value, int N, double *&result);
+int compact(float *index, int *value, int N, float *&result);
 
 template <typename T>
 void allocate_and_copy(T *dev_index, int *dev_value, int *dev_output, int N, int arrSize, T *dev_result_index);
@@ -42,6 +42,6 @@ int compact_non_native(int N);
 
 int compact_native(int N);
 
-//__global__ void dummy(int *dev_ibead_lj_nat, int *dev_jbead_lj_nat, int *dev_itype_lj_nat, int *dev_jtype_lj_nat, double3 *dev_unc_pos, double *dev_lj_nat_pdb_dist, int *&dev_value, int boxl, int N, int nbead);
+//__global__ void dummy(int *dev_ibead_lj_nat, int *dev_jbead_lj_nat, int *dev_itype_lj_nat, int *dev_jtype_lj_nat, float3 *dev_unc_pos, float *dev_lj_nat_pdb_dist, int *&dev_value, int boxl, int N, int nbead);
 
 #endif
